* use asInt64 instead of asInt to evaluate an Integer expression
git-svn-id: trunk@45052 -

+++ compiler/scanner.pas
@@ -1329,12 +1329,12 @@ type
 
   function texprvalue.isBoolean: Boolean;
     var
-      i: integer;
+      i: int64;
     begin
       result:=is_boolean(def);
       if not result and is_integer(def) then
         begin
-          i:=asInt;
+          i:=asInt64;
           result:=(i=0)or(i=1);
         end;
     end;
